split activation related functions to its own module
